refactor some code, add extra doip_send function
